requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
yfinance>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.22.0
pyyaml>=6.0
typing-extensions>=4.5.0
numpy>=1.24.0
plotly>=5.13.0